Moved aggregate tvl api related test to other appropriate file Fixed linting issues
