+++ packages/backend/package.json
@@ -57,9 +57,9 @@
     "@types/basic-auth": "^1.1.3",
     "@types/deep-diff": "^1.0.2",
     "@types/koa": "^2.13.5",
+    "@types/koa__router": "^12.0.0",
     "@types/koa-conditional-get": "^2.0.0",
     "@types/koa-etag": "^3.0.0",
-    "@types/koa__router": "^12.0.0",
     "@types/lodash": "^4.14.191",
     "@types/node-fetch": "^2.6.2",
     "@types/pg": "8.6.5",
